% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dataCleaning.R
\name{dataCleaningDeliveries}
\alias{dataCleaningDeliveries}
\title{dataCleaningDeliveries}
\usage{
dataCleaningDeliveries(x)
}
\arguments{
\item{x}{dataframe of master deliveries}
}
\value{
cleaned up dataframe same format
}
\description{
Additional data cleaning missed in pre processing files for refinery deliveries
}
