% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dataCleaning.R
\name{dataCleaningProduction}
\alias{dataCleaningProduction}
\title{dataCleaningProduction}
\usage{
dataCleaningProduction(x)
}
\arguments{
\item{x}{dataframe of masterProd}
}
\value{
cleaned up dataframe same format
}
\description{
Additional data cleaning missed in pre processing files for refinery production
}
