% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dataCleaning.R
\name{dataCleaningReceipts}
\alias{dataCleaningReceipts}
\title{dataCleaningReceipts}
\usage{
dataCleaningReceipts(x)
}
\arguments{
\item{x}{dataframe of master receipts}
}
\value{
cleaned up dataframe same format
}
\description{
Additional data cleaning missed in pre processing files for refinery receipts
}
