% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{masterDeliveries}
\alias{masterDeliveries}
\title{dataset: aggregated delivery data for refineries}
\format{
data frame
}
\usage{
masterDeliveries
}
\description{
all products and companies delivered from refinery
}
\keyword{datasets}
