% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{masterProd}
\alias{masterProd}
\title{dataset: aggregated production data for refineries}
\format{
data frame
}
\usage{
masterProd
}
\description{
production data from refineries in usgc
}
\keyword{datasets}
