% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{masterReceipts}
\alias{masterReceipts}
\title{dataset: aggregated receipts data for refineries}
\format{
data frame
}
\usage{
masterReceipts
}
\description{
all products and companies sending to refinery
}
\keyword{datasets}
