% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refShape}
\alias{refShape}
\title{dataset: shapefile for refineries and capacities in texas}
\format{
sf
}
\usage{
refShape
}
\description{
capacities and locations for refineries in texas
}
\keyword{datasets}
